jobdesk/master: progress, commons: improve
- progress in `jobdesk/master`
  - clean pom.xml dependencies (remove spring-boot-web module)
  - switch from boolean primitive type to Boolean class for nullable SQL fields
  - "compress" the previous Batch implementations into something
    hopefully simpler. Still have problem to build clean Integration
    Test Suite though...
- add some improvements to `commons-sql` stuff (show SQLException
  original exception when occurred in a batch process, add two
  "factories" to easy build JDBC readers / jOOQ writers)
- add `commons-config` module to share common Spring Boot and Logging settings
- add support for an embedded ElasticSearch node in `commons-test` module
- the majority of unit/integration tests are based on
  SimpleTestApplication defined in `commons-test`, which forces to store
  the SpringBatch metadata into an in-memory database.

+++ jobdesk/master/src/main/java/ch/alv/batches/master/to/jobdesk/model/JobdeskJobContact.java
@@ -5,7 +5,8 @@ package ch.alv.batches.master.to.jobdesk.model;
  */
 public class JobdeskJobContact {
 
-    private int gender;
+    // FIXME Enumeration? Constants at least?
+    private Short gender;
 
     private String firstName;
 
@@ -18,7 +19,7 @@ public class JobdeskJobContact {
     public JobdeskJobContact() {
     }
 
-    public JobdeskJobContact(int gender, String firstName, String lastName, String phone, String email) {
+    public JobdeskJobContact(Short gender, String firstName, String lastName, String phone, String email) {
         this.gender = gender;
         this.firstName = firstName;
         this.lastName = lastName;
@@ -26,11 +27,11 @@ public class JobdeskJobContact {
         this.email = email;
     }
 
-    public int getGender() {
+    public Short getGender() {
         return gender;
     }
 
-    public void setGender(int gender) {
+    public void setGender(Short gender) {
         this.gender = gender;
     }
